# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "gRPC endpoints docs",
    "description": "",
    "version": "1.0.0"
  },
  "host": "",
...
  "paths": {
...
    "/users/{userId}/payments": {
      "get": {
        "tags": [
          "GRPC.User.Payment"
        ],
        "description": "Endpoint to get payments via grpc",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id of specific user."
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
    "/payments": {
      "get": {
        "tags": [
          "Payment"
        ],
        "description": "Endpoint to get PAYMENTS",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
    "/payments/{paymentId}/Pay": {
      "get": {
        "tags": [
          "Payment"
        ],
        "description": "Endpoint to iniliatize specific payment",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Payment identification"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
    "/payments/result": {
      "get": {
        "tags": [
          "Payment"
        ],
        "description": "Endpoint to get result of payment",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "JWTAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}